third_party/blink/web_tests/external/wpt/editing/run/underline_1-1000-expected.txt

This is a testharness.js-based test.
Found 169 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["underline",""]] "foo[]bar" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""]] "foo[]bar" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["underline",""]] "<p>[foo</p> <p>bar]</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><span style=\\"text-decoration:underline\\">foo</span></p> <p><span style=\\"text-decoration:underline\\">bar</span></p>" but got "<p><span style=\\"text-decoration-line:underline\\">foo</span></p> <p><span style=\\"text-decoration-line:underline\\">bar</span></p>"
[FAIL] [["stylewithcss","true"],["underline",""]] "<p>[foo</p> <p>bar]</p>" queryCommandState("stylewithcss") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","true"],["underline",""]] "<p>[foo</p> <p>bar]</p>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "<p>[foo</p> <p>bar]</p>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["underline",""]] "<p>[foo</p> <p>bar]</p>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "<p>[foo</p> <p>bar]</p>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["underline",""]] "<span>[foo</span> <span>bar]</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"text-decoration:underline\\"><span>foo</span> <span>bar</span></span>" but got "<span style=\\"text-decoration-line:underline\\"><span>foo</span> <span>bar</span></span>"
[FAIL] [["stylewithcss","true"],["underline",""]] "<span>[foo</span> <span>bar]</span>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "<span>[foo</span> <span>bar]</span>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["underline",""]] "<span>[foo</span> <span>bar]</span>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "<span>[foo</span> <span>bar]</span>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["underline",""]] "<p>[foo</p><p> <span>bar</span> </p><p>baz]</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><span style=\\"text-decoration:underline\\">foo</span></p><p> <span style=\\"text-decoration:underline\\">bar</span> </p><p><span style=\\"text-decoration:underline\\">baz</span></p>" but got "<p><span style=\\"text-decoration-line:underline\\">foo</span></p><p> <span style=\\"text-decoration-line:underline\\"><span>bar</span> </span></p><p><span style=\\"text-decoration-line:underline\\">baz</span></p>"
[FAIL] [["stylewithcss","true"],["underline",""]] "<p>[foo</p><p> <span>bar</span> </p><p>baz]</p>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "<p>[foo</p><p> <span>bar</span> </p><p>baz]</p>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["underline",""]] "<p>[foo</p><p> <span>bar</span> </p><p>baz]</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><u>foo</u></p><p> <u><span>bar</span></u> </p><p><u>baz</u></p>" but got "<p><u>foo</u></p><p> <u><span>bar</span> </u></p><p><u>baz</u></p>"
[FAIL] [["stylewithcss","false"],["underline",""]] "<p>[foo</p><p> <span>bar</span> </p><p>baz]</p>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "<p>[foo</p><p> <span>bar</span> </p><p>baz]</p>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["underline",""]] "<p>[foo<p><br><p>bar]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><span style=\\"text-decoration:underline\\">foo</span></p><p><span style=\\"text-decoration:underline\\"><br></span></p><p><span style=\\"text-decoration:underline\\">bar</span></p>" but got "<p><span style=\\"text-decoration-line:underline\\">foo</span></p><p><span style=\\"text-decoration-line:underline\\"><br></span></p><p><span style=\\"text-decoration-line:underline\\">bar</span></p>"
[FAIL] [["stylewithcss","true"],["underline",""]] "<p>[foo<p><br><p>bar]" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "<p>[foo<p><br><p>bar]" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["underline",""]] "<p>[foo<p><br><p>bar]" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "<p>[foo<p><br><p>bar]" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["underline",""]] "<b>foo[]bar</b>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""]] "<b>foo[]bar</b>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["underline",""]] "<i>foo[]bar</i>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""]] "<i>foo[]bar</i>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["underline",""]] "<span>foo</span>{}<span>bar</span>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""]] "<span>foo</span>{}<span>bar</span>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["underline",""]] "<span>foo[</span><span>]bar</span>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""]] "<span>foo[</span><span>]bar</span>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo[bar]baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span style=\\"text-decoration:underline\\">bar</span>baz" but got "foo<span style=\\"text-decoration-line:underline\\">bar</span>baz"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo[bar]baz" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo[bar]baz" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["underline",""]] "foo[bar]baz" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "foo[bar]baz" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo[bar<b>baz]qoz</b>quz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span style=\\"text-decoration:underline\\">bar</span><b><span style=\\"text-decoration:underline\\">baz</span>qoz</b>quz" but got "foo<span style=\\"text-decoration-line:underline\\">bar</span><b><span style=\\"text-decoration-line:underline\\">baz</span>qoz</b>quz"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo[bar<b>baz]qoz</b>quz" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo[bar<b>baz]qoz</b>quz" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["underline",""]] "foo[bar<b>baz]qoz</b>quz" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "foo[bar<b>baz]qoz</b>quz" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo[bar<i>baz]qoz</i>quz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span style=\\"text-decoration:underline\\">bar</span><i><span style=\\"text-decoration:underline\\">baz</span>qoz</i>quz" but got "foo<span style=\\"text-decoration-line:underline\\">bar</span><i><span style=\\"text-decoration-line:underline\\">baz</span>qoz</i>quz"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo[bar<i>baz]qoz</i>quz" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo[bar<i>baz]qoz</i>quz" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["underline",""]] "foo[bar<i>baz]qoz</i>quz" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "foo[bar<i>baz]qoz</i>quz" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["underline",""]] "{<p><p> <p>foo</p>}" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p></p><p> </p><p><span style=\\"text-decoration:underline\\">foo</span></p>" but got "<p></p><p> </p><p><span style=\\"text-decoration-line:underline\\">foo</span></p>"
[FAIL] [["stylewithcss","true"],["underline",""]] "{<p><p> <p>foo</p>}" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "{<p><p> <p>foo</p>}" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["underline",""]] "{<p><p> <p>foo</p>}" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "{<p><p> <p>foo</p>}" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["underline",""]] "<table><tbody><tr><td>foo<td>b[a]r<td>baz</table>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<table><tbody><tr><td>foo</td><td>b<span style=\\"text-decoration:underline\\">a</span>r</td><td>baz</td></tr></tbody></table>" but got "<table><tbody><tr><td>foo</td><td>b<span style=\\"text-decoration-line:underline\\">a</span>r</td><td>baz</td></tr></tbody></table>"
[FAIL] [["stylewithcss","true"],["underline",""]] "<table><tbody><tr><td>foo<td>b[a]r<td>baz</table>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "<table><tbody><tr><td>foo<td>b[a]r<td>baz</table>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["underline",""]] "<table><tbody><tr><td>foo<td>b[a]r<td>baz</table>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "<table><tbody><tr><td>foo<td>b[a]r<td>baz</table>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["underline",""]] "<table><tbody><tr data-start=1 data-end=2><td>foo<td>bar<td>baz</table>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<table><tbody><tr><td>foo</td><td><span style=\\"text-decoration:underline\\">bar</span></td><td>baz</td></tr></tbody></table>" but got "<table><tbody><tr><td>foo</td><td><span style=\\"text-decoration-line:underline\\">bar</span></td><td>baz</td></tr></tbody></table>"
[FAIL] [["stylewithcss","true"],["underline",""]] "<table><tbody><tr data-start=1 data-end=2><td>foo<td>bar<td>baz</table>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "<table><tbody><tr data-start=1 data-end=2><td>foo<td>bar<td>baz</table>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["underline",""]] "<table><tbody><tr data-start=1 data-end=2><td>foo<td>bar<td>baz</table>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "<table><tbody><tr data-start=1 data-end=2><td>foo<td>bar<td>baz</table>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["underline",""]] "<table><tbody><tr data-start=0 data-end=2><td>foo<td>bar<td>baz</table>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<table><tbody><tr><td><span style=\\"text-decoration:underline\\">foo</span></td><td><span style=\\"text-decoration:underline\\">bar</span></td><td>baz</td></tr></tbody></table>" but got "<table><tbody><tr><td><span style=\\"text-decoration-line:underline\\">foo</span></td><td><span style=\\"text-decoration-line:underline\\">bar</span></td><td>baz</td></tr></tbody></table>"
[FAIL] [["stylewithcss","true"],["underline",""]] "<table><tbody><tr data-start=0 data-end=2><td>foo<td>bar<td>baz</table>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "<table><tbody><tr data-start=0 data-end=2><td>foo<td>bar<td>baz</table>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["underline",""]] "<table><tbody><tr data-start=0 data-end=2><td>foo<td>bar<td>baz</table>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "<table><tbody><tr data-start=0 data-end=2><td>foo<td>bar<td>baz</table>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["underline",""]] "<table><tbody data-start=0 data-end=1><tr><td>foo<td>bar<td>baz</table>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<table><tbody><tr><td><span style=\\"text-decoration:underline\\">foo</span></td><td><span style=\\"text-decoration:underline\\">bar</span></td><td><span style=\\"text-decoration:underline\\">baz</span></td></tr></tbody></table>" but got "<table><tbody><tr><td><span style=\\"text-decoration-line:underline\\">foo</span></td><td><span style=\\"text-decoration-line:underline\\">bar</span></td><td><span style=\\"text-decoration-line:underline\\">baz</span></td></tr></tbody></table>"
[FAIL] [["stylewithcss","true"],["underline",""]] "<table><tbody data-start=0 data-end=1><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "<table><tbody data-start=0 data-end=1><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["underline",""]] "<table><tbody data-start=0 data-end=1><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "<table><tbody data-start=0 data-end=1><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["underline",""]] "<table data-start=0 data-end=1><tbody><tr><td>foo<td>bar<td>baz</table>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<table><tbody><tr><td><span style=\\"text-decoration:underline\\">foo</span></td><td><span style=\\"text-decoration:underline\\">bar</span></td><td><span style=\\"text-decoration:underline\\">baz</span></td></tr></tbody></table>" but got "<table><tbody><tr><td><span style=\\"text-decoration-line:underline\\">foo</span></td><td><span style=\\"text-decoration-line:underline\\">bar</span></td><td><span style=\\"text-decoration-line:underline\\">baz</span></td></tr></tbody></table>"
[FAIL] [["stylewithcss","true"],["underline",""]] "<table data-start=0 data-end=1><tbody><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "<table data-start=0 data-end=1><tbody><tr><td>foo<td>bar<td>baz</table>" queryCommandState("underline") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["underline",""]] "<table data-start=0 data-end=1><tbody><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "<table data-start=0 data-end=1><tbody><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "<table data-start=0 data-end=1><tbody><tr><td>foo<td>bar<td>baz</table>" queryCommandState("underline") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["underline",""]] "<table data-start=0 data-end=1><tbody><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "{<table><tr><td>foo<td>bar<td>baz</table>}" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<table><tbody><tr><td><span style=\\"text-decoration:underline\\">foo</span></td><td><span style=\\"text-decoration:underline\\">bar</span></td><td><span style=\\"text-decoration:underline\\">baz</span></td></tr></tbody></table>" but got "<table><tbody><tr><td><span style=\\"text-decoration-line:underline\\">foo</span></td><td><span style=\\"text-decoration-line:underline\\">bar</span></td><td><span style=\\"text-decoration-line:underline\\">baz</span></td></tr></tbody></table>"
[FAIL] [["stylewithcss","true"],["underline",""]] "{<table><tr><td>foo<td>bar<td>baz</table>}" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "{<table><tr><td>foo<td>bar<td>baz</table>}" queryCommandState("underline") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["underline",""]] "{<table><tr><td>foo<td>bar<td>baz</table>}" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "{<table><tr><td>foo<td>bar<td>baz</table>}" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "{<table><tr><td>foo<td>bar<td>baz</table>}" queryCommandState("underline") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["underline",""]] "{<table><tr><td>foo<td>bar<td>baz</table>}" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo<span style=\\"text-decoration: underline\\">[bar]</span>baz" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo<span style=\\"text-decoration: underline\\">[bar]</span>baz" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "foo<span style=\\"text-decoration: underline\\">[bar]</span>baz" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["underline",""]] "foo<span style=\\"text-decoration: underline\\">[bar]</span>baz" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "<u>foo[bar]baz</u>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"text-decoration:underline\\">foo</span>bar<span style=\\"text-decoration:underline\\">baz</span>" but got "<u>foo</u>bar<u>baz</u>"
[FAIL] [["stylewithcss","true"],["underline",""]] "<u>foo[bar]baz</u>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["underline",""]] "<u>foo[bar]baz</u>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "<u>foo[bar]baz</u>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["underline",""]] "<u>foo[bar]baz</u>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "<u>foo[b<span style=\\"color:blue\\">ar]ba</span>z</u>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"text-decoration:underline\\">foo</span>b<span style=\\"color:rgb(0, 0, 255)\\">ar<span style=\\"text-decoration:underline\\">ba</span></span><span style=\\"text-decoration:underline\\">z</span>" but got "<u>foo</u>b<span style=\\"color:rgb(0, 0, 255)\\">ar<span style=\\"text-decoration-line:underline\\">ba</span></span><span style=\\"text-decoration-line:underline\\">z</span>"
[FAIL] [["stylewithcss","true"],["underline",""]] "<u>foo[b<span style=\\"color:blue\\">ar]ba</span>z</u>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["underline",""]] "<u>foo[b<span style=\\"color:blue\\">ar]ba</span>z</u>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "<u>foo[b<span style=\\"color:blue\\">ar]ba</span>z</u>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["underline",""]] "<u>foo[b<span style=\\"color:blue\\">ar]ba</span>z</u>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "<u>foo[b<span style=\\"color:blue\\" id=foo>ar]ba</span>z</u>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"text-decoration:underline\\">foo</span>b<span style=\\"color:rgb(0, 0, 255)\\" id=\\"foo\\">ar<span style=\\"text-decoration:underline\\">ba</span></span><span style=\\"text-decoration:underline\\">z</span>" but got "<u>foo</u>b<span style=\\"color:rgb(0, 0, 255)\\" id=\\"foo\\">ar<span style=\\"text-decoration-line:underline\\">ba</span></span><span style=\\"text-decoration-line:underline\\">z</span>"
[FAIL] [["stylewithcss","true"],["underline",""]] "<u>foo[b<span style=\\"color:blue\\" id=foo>ar]ba</span>z</u>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["underline",""]] "<u>foo[b<span style=\\"color:blue\\" id=foo>ar]ba</span>z</u>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "<u>foo[b<span style=\\"color:blue\\" id=foo>ar]ba</span>z</u>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["underline",""]] "<u>foo[b<span style=\\"color:blue\\" id=foo>ar]ba</span>z</u>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "<u>foo[b<span style=\\"font-size:3em\\">ar]ba</span>z</u>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"text-decoration:underline\\">foo</span>b<span style=\\"font-size:3em\\">ar<span style=\\"text-decoration:underline\\">ba</span></span><span style=\\"text-decoration:underline\\">z</span>" but got "<u>foo</u>b<span style=\\"font-size:3em\\">ar<span style=\\"text-decoration-line:underline\\">ba</span></span><span style=\\"text-decoration-line:underline\\">z</span>"
[FAIL] [["stylewithcss","true"],["underline",""]] "<u>foo[b<span style=\\"font-size:3em\\">ar]ba</span>z</u>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["underline",""]] "<u>foo[b<span style=\\"font-size:3em\\">ar]ba</span>z</u>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "<u>foo[b<span style=\\"font-size:3em\\">ar]ba</span>z</u>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["underline",""]] "<u>foo[b<span style=\\"font-size:3em\\">ar]ba</span>z</u>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "<u>foo[b<i>ar]ba</i>z</u>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"text-decoration:underline\\">foo</span>b<i>ar<span style=\\"text-decoration:underline\\">ba</span></i><span style=\\"text-decoration:underline\\">z</span>" but got "<u>foo</u>b<i style=\\"\\">ar<span style=\\"text-decoration-line:underline\\">ba</span></i><span style=\\"text-decoration-line:underline\\">z</span>"
[FAIL] [["stylewithcss","true"],["underline",""]] "<u>foo[b<i>ar]ba</i>z</u>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["underline",""]] "<u>foo[b<i>ar]ba</i>z</u>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "<u>foo[b<i>ar]ba</i>z</u>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<u>foo</u>b<i>ar<u>ba</u></i><u>z</u>" but got "<u>foo</u>b<i style=\\"\\">ar<u>ba</u></i><u>z</u>"
[FAIL] [["stylewithcss","false"],["underline",""]] "<u>foo[b<i>ar]ba</i>z</u>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["underline",""]] "<u>foo[b<i>ar]ba</i>z</u>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "<p style=\\"text-decoration: underline\\">foo[bar]baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><span style=\\"text-decoration:underline\\">foo</span>bar<span style=\\"text-decoration:underline\\">baz</span></p>" but got "<p style=\\"\\"><span style=\\"text-decoration-line:underline\\">foo</span>bar<span style=\\"text-decoration-line:underline\\">baz</span></p>"
[FAIL] [["stylewithcss","true"],["underline",""]] "<p style=\\"text-decoration: underline\\">foo[bar]baz</p>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["underline",""]] "<p style=\\"text-decoration: underline\\">foo[bar]baz</p>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "<p style=\\"text-decoration: underline\\">foo[bar]baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><u>foo</u>bar<u>baz</u></p>" but got "<p style=\\"\\"><span style=\\"text-decoration-line:underline\\">foo</span>bar<u>baz</u></p>"
[FAIL] [["stylewithcss","false"],["underline",""]] "<p style=\\"text-decoration: underline\\">foo[bar]baz</p>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["underline",""]] "<p style=\\"text-decoration: underline\\">foo[bar]baz</p>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo<s>[bar]</s>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span style=\\"text-decoration:underline line-through\\">bar</span>baz" but got "foo<span style=\\"text-decoration-line:underline line-through\\">bar</span>baz"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo<s>[bar]</s>baz" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo<s>[bar]</s>baz" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["underline",""]] "foo<s>[bar]</s>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<s><u>bar</u></s>baz" but got "foo<u><strike>bar</strike></u>baz"
[FAIL] [["stylewithcss","false"],["underline",""]] "foo<s>[bar]</s>baz" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "foo<s>[bar]</s>baz" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo<span style=\\"text-decoration: line-through\\">[bar]</span>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span style=\\"text-decoration:underline line-through\\">bar</span>baz" but got "foo<span style=\\"text-decoration-line:underline line-through\\">bar</span>baz"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo<span style=\\"text-decoration: line-through\\">[bar]</span>baz" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo<span style=\\"text-decoration: line-through\\">[bar]</span>baz" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["underline",""]] "foo<span style=\\"text-decoration: line-through\\">[bar]</span>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<u><span style=\\"text-decoration:line-through\\">bar</span></u>baz" but got "foo<u><strike>bar</strike></u>baz"
[FAIL] [["stylewithcss","false"],["underline",""]] "foo<span style=\\"text-decoration: line-through\\">[bar]</span>baz" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "foo<span style=\\"text-decoration: line-through\\">[bar]</span>baz" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["underline",""]] "<s>foo[bar]baz</s>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<s>foo<span style=\\"text-decoration:underline\\">bar</span>baz</s>" but got "<s>foo<span style=\\"text-decoration-line:underline\\">bar</span>baz</s>"
[FAIL] [["stylewithcss","true"],["underline",""]] "<s>foo[bar]baz</s>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "<s>foo[bar]baz</s>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["underline",""]] "<s>foo[bar]baz</s>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "<s>foo[bar]baz</s>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["underline",""]] "<s>foo[b<span style=\\"color:blue\\">ar]ba</span>z</s>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<s>foo<span style=\\"text-decoration:underline\\">b</span><span style=\\"color:rgb(0, 0, 255)\\"><span style=\\"text-decoration:underline\\">ar</span>ba</span>z</s>" but got "<s>foo<span style=\\"text-decoration-line:underline\\">b</span><span style=\\"color:rgb(0, 0, 255)\\"><span style=\\"text-decoration-line:underline\\">ar</span>ba</span>z</s>"
[FAIL] [["stylewithcss","true"],["underline",""]] "<s>foo[b<span style=\\"color:blue\\">ar]ba</span>z</s>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "<s>foo[b<span style=\\"color:blue\\">ar]ba</span>z</s>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["underline",""]] "<s>foo[b<span style=\\"color:blue\\">ar]ba</span>z</s>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "<s>foo[b<span style=\\"color:blue\\">ar]ba</span>z</s>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["underline",""]] "<s>foo[b<span style=\\"color:blue\\" id=foo>ar]ba</span>z</s>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<s>foo<span style=\\"text-decoration:underline\\">b</span><span style=\\"color:rgb(0, 0, 255)\\" id=\\"foo\\"><span style=\\"text-decoration:underline\\">ar</span>ba</span>z</s>" but got "<s>foo<span style=\\"text-decoration-line:underline\\">b</span><span style=\\"color:rgb(0, 0, 255)\\" id=\\"foo\\"><span style=\\"text-decoration-line:underline\\">ar</span>ba</span>z</s>"
[FAIL] [["stylewithcss","true"],["underline",""]] "<s>foo[b<span style=\\"color:blue\\" id=foo>ar]ba</span>z</s>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "<s>foo[b<span style=\\"color:blue\\" id=foo>ar]ba</span>z</s>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["underline",""]] "<s>foo[b<span style=\\"color:blue\\" id=foo>ar]ba</span>z</s>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "<s>foo[b<span style=\\"color:blue\\" id=foo>ar]ba</span>z</s>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["underline",""]] "<s>foo[b<span style=\\"font-size:3em\\">ar]ba</span>z</s>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<s>foo<span style=\\"text-decoration:underline\\">b</span><span style=\\"font-size:3em\\"><span style=\\"text-decoration:underline\\">ar</span>ba</span>z</s>" but got "<s>foo<span style=\\"text-decoration-line:underline\\">b</span><span style=\\"font-size:3em\\"><span style=\\"text-decoration-line:underline\\">ar</span>ba</span>z</s>"
[FAIL] [["stylewithcss","true"],["underline",""]] "<s>foo[b<span style=\\"font-size:3em\\">ar]ba</span>z</s>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "<s>foo[b<span style=\\"font-size:3em\\">ar]ba</span>z</s>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["underline",""]] "<s>foo[b<span style=\\"font-size:3em\\">ar]ba</span>z</s>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "<s>foo[b<span style=\\"font-size:3em\\">ar]ba</span>z</s>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["underline",""]] "<s>foo[b<i>ar]ba</i>z</s>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<s>foo<span style=\\"text-decoration:underline\\">b</span><i><span style=\\"text-decoration:underline\\">ar</span>ba</i>z</s>" but got "<s>foo<span style=\\"text-decoration-line:underline\\">b</span><i><span style=\\"text-decoration-line:underline\\">ar</span>ba</i>z</s>"
[FAIL] [["stylewithcss","true"],["underline",""]] "<s>foo[b<i>ar]ba</i>z</s>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "<s>foo[b<i>ar]ba</i>z</s>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["underline",""]] "<s>foo[b<i>ar]ba</i>z</s>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "<s>foo[b<i>ar]ba</i>z</s>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["underline",""]] "<p style=\\"text-decoration: line-through\\">foo[bar]baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p style=\\"text-decoration:line-through\\">foo<span style=\\"text-decoration:underline\\">bar</span>baz</p>" but got "<p style=\\"\\"><span style=\\"text-decoration-line:line-through\\">foo</span><span style=\\"text-decoration-line:underline line-through\\">bar</span><span style=\\"text-decoration-line:line-through\\">baz</span></p>"
[FAIL] [["stylewithcss","true"],["underline",""]] "<p style=\\"text-decoration: line-through\\">foo[bar]baz</p>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "<p style=\\"text-decoration: line-through\\">foo[bar]baz</p>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["underline",""]] "<p style=\\"text-decoration: line-through\\">foo[bar]baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p style=\\"text-decoration:line-through\\">foo<u>bar</u>baz</p>" but got "<p style=\\"\\"><span style=\\"text-decoration-line:line-through\\">foo</span><u><strike>bar</strike></u><strike>baz</strike></p>"
[FAIL] [["stylewithcss","false"],["underline",""]] "<p style=\\"text-decoration: line-through\\">foo[bar]baz</p>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "<p style=\\"text-decoration: line-through\\">foo[bar]baz</p>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo<strike>[bar]</strike>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span style=\\"text-decoration:underline line-through\\">bar</span>baz" but got "foo<span style=\\"text-decoration-line:underline line-through\\">bar</span>baz"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo<strike>[bar]</strike>baz" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo<strike>[bar]</strike>baz" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["underline",""]] "foo<strike>[bar]</strike>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<strike><u>bar</u></strike>baz" but got "foo<u><strike>bar</strike></u>baz"
Harness: the test ran to completion.
